requirements.txt
numpy<2
pandas
scikit-learn
tensorflow
streamlit
matplotlib
plotly
google-generativeai
python-dotenv